+++ HelloGame/HelloGame.Tests/TestX.cs
@@ -4,12 +4,12 @@ namespace HelloGame.Tests
 {
     static class TestX
     {
-        public static void AssertAreAlmostSame(double a, double b)
+        public static void AssertAreAlmostSame(decimal a, decimal b)
         {
             Assert.IsTrue(MathX.IsAlmostSame(a, b));
         }
 
-        public static void AssertIsAlmostZero(double a)
+        public static void AssertIsAlmostZero(decimal a)
         {
             AssertAreAlmostSame(a, 0);
         }
